* 'master' of https://github.com/binscot/hanghae-todaydiary:
  다이어리 삭제 구현 완료
  다이어리 삭제
  comment 수정 시간 데이터 수정
  다이어리 삭제
  imageurl service 수정 2
  imageurl service 수정 2
  명칭 수정 후 커밋.
  이미지url service 수정
  이미지url service 설정
  image url 리스트 수정
  image url 리스트 수정
  image url 리스트 수정
  /api/user을 추가하여 커밋~!!!
  jwt 업데이트.

# Conflicts:
#	src/main/java/com/example/todaydiary/diary/ImageUrl.java
#	src/main/java/com/example/todaydiary/diary/ImageUrlRepository.java

+++ src/main/java/com/example/todaydiary/diary/DiaryService.java
@@ -4,36 +4,43 @@ import com.example.todaydiary.security.UserDetailsImpl;
 import com.example.todaydiary.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @RequiredArgsConstructor
 @Service
 public class DiaryService {
 
     private final DiaryRepository diaryRepository;
+    private final ImageUrlRepository imageUrlRepository;
 
 
     //글 작성
     @Transactional
-    public Diary createDiary(DiaryRequestDto requestDto, User user) {
+    public Diary createDiary(
+            DiaryRequestDto requestDto,
+            User user) {
 
-        //
+        List<ImageUrl> imageUrlList1 = new ArrayList<>();
+        for (ImageUrl imageUrls : requestDto.getImageUrlList()) {
+            imageUrlList1.add(imageUrlRepository.save(imageUrls)
+            );
+        }
+
         String content = requestDto.getContent();
-        if(requestDto.getContent()==null){
+        if (requestDto.getContent() == null) {
             throw new IllegalArgumentException("내용을 입력해주세요.");
         }
-        if(content.length() >1000){
+        if (content.length() > 1000) {
             throw new IllegalArgumentException("1000자 이하로 입력해주세요.");
         }
 
-        Diary diary = new Diary(requestDto, user);
-        diaryRepository.save(diary);
-        return diary;
+        Diary diary = new Diary(requestDto, user, imageUrlList1);
+
+        return diaryRepository.save(diary);
     }
 
 
@@ -42,32 +49,37 @@ public class DiaryService {
     public Diary updateDiary(
             Long id,
             DiaryRequestDto requestDto,
-            UserDetailsImpl userDetails)
-    {
-       Diary diary = diaryRepository.findById(id).orElseThrow(
+            UserDetailsImpl userDetails) {
+        Diary diary = diaryRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("일기가 존재하지 않습니다.")
         );
-       User user = diary.getUser();
-       if(userDetails.getUser() != user){
-           throw new IllegalArgumentException("작성자만 수정할 수 있습니다.");
-       }
-       if(requestDto.getContent()==null){
-           throw new IllegalArgumentException("내용을 입력해주세요.");
-       }
-       if(requestDto.getContent().length() > 1000){
-           throw new IllegalArgumentException ("1000자 이하로 입력해주세요.");
-       }
+        User user = diary.getUser();
+        if (userDetails.getUser() != user) {
+            throw new IllegalArgumentException("작성자만 수정할 수 있습니다.");
+        }
+        if (requestDto.getContent() == null) {
+            throw new IllegalArgumentException("내용을 입력해주세요.");
+        }
+        if (requestDto.getContent().length() > 1000) {
+            throw new IllegalArgumentException("1000자 이하로 입력해주세요.");
+        }
         diary.updateDiary(requestDto);
-       diaryRepository.save(diary);
+        diaryRepository.save(diary);
         return diary;
     }
 
 
     //삭제
     @Transactional
-    public Long deleteDiary(Long diaryId) {
-        Diary diary = diaryRepository.findById(diaryId)
-                        .orElseThrow(()->new IllegalArgumentException("일기가 없습니다."));
+    public Long deleteDiary(Long diaryId, UserDetailsImpl userDetails) {
+        Diary diary = diaryRepository.findById(diaryId).orElseThrow(
+                () -> new IllegalArgumentException("일기가 존재하지 않습니다.")
+        );
+        User user = diary.getUser();
+        Long deleteId = user.getId();
+        if (!Objects.equals(userDetails.getUser().getId(), deleteId)) {
+            throw new IllegalArgumentException("작성자만 수정할 수 있습니다.");
+        }
             diaryRepository.deleteById(diaryId);
             return diaryId;
         }
